Add --simulate and --keep_path_element options

+++ script.py
@@ -2,7 +2,12 @@
 
 """
 Usage:
-    {self_filename} LIST_FILE OUT_DIR [--options=<ffmpeg_options>]
+    {self_filename} LIST_FILE OUT_DIR [--options=<ffmpeg_options>] [options]
+
+Options:
+    --simulate                      Do not actually convert files, only print commands that would be executed
+    --keep_path_elements=<n>        [default: 2]
+
 """
 
 from docopt import docopt
@@ -29,8 +34,9 @@ with list_file.open() as f:
         if line == '':
             continue
         file = Path(line)                                                           # /media/bruno/BU2/FAE_20231026_002/NINJAV_S001_S001_T002.MOV
-        out_file = Path(*file.parts[-2:])                                           # FAE_20231026_002/NINJAV_S001_S001_T002.MOV
-        out_file = Path(out_dir / out_file.parent / Path(out_file.stem + '.mp4'))   # out/DD2/FAE_20231026_002/NINJAV_S001_S001_T002.mp4
+        nb_path_elements = int(args['--keep_path_elements'])
+        out_file = Path(*file.parts[-nb_path_elements:])                            # DD1/FAE_20231026_002/NINJAV_S001_S001_T002.MOV (si 3 elements)
+        out_file = Path(out_dir / out_file.parent / Path(out_file.stem + '.mp4'))   # /path/to/output/DD1/FAE_20231026_002/NINJAV_S001_S001_T002.mp4
         print(f"{line}  ->  {out_file}")
         tmp_file = f"tmp/{out_file.name}"                                           # tmp/NINJAV_S001_S001_T002.mp4
         os.system(f"rm -rf tmp/*")
@@ -38,8 +44,10 @@ with list_file.open() as f:
         print("Duration: {0:.1f}m".format(float(duration.decode('utf8')) / 60))
         command = f"ffpb -i {line} {ffmpeg_options} {tmp_file}"
         print(command)
-        errors = subprocess.call(command, shell=True)
-        assert not errors
-        out_file.parent.mkdir(parents=True, exist_ok=True)
-        shutil.move(tmp_file, out_file.parent)
+        # Convertion
+        if not args['--simulate']:
+            errors = subprocess.call(command, shell=True)
+            assert not errors
+            out_file.parent.mkdir(parents=True, exist_ok=True)
+            shutil.move(tmp_file, out_file.parent)
         print("\n\n")
